// EscalafonFrontend/src/app/interfaces/v-compensacion.ts
export interface VCompensacion {
    /**
     * @type { number } - iLegId
     */
    iLegId?: number;

    /**
     * @type { number } - iTipoDocId
     */
    iTipoDocId?: number;

    /**
     * @type { number } - iArchId
     */
    iArchId?: number;

    /**
     * @type { number } - iRegLabId
     */
    iRegLabId?: number;

    /**
     * @type { number } - iCompAccId
     */
    iCompAccId?: number;

    /**
     * @type { number } - iCompMotAccId
     */
    iCompMotAccId?: number;

    /**
     * @type { number } - iComTipFallId
     */
    iComTipFallId?: number;

    /**
     * @type { number } - iComTipMonId
     */
    iComTipMonId?: number;

    /**
     * @type { string } - cAsigIncNumeroDocumento
     */
    cAsigIncNumeroDocumento?: string;

    /**
     * @type { Date } - dtAsigIncFechaDocumento
     */
    dtAsigIncFechaDocumento?: Date;

    /**
     * @type { string } - cAsigIncDerechoHabienteFallecido
     */
    cAsigIncDerechoHabienteFallecido?: string;

    /**
     * @type { string } - cAsigIncDerechoSubsidiado
     */
    cAsigIncDerechoSubsidiado?: string;

    /**
     * @type { number } - iCompTipPagId
     */
    iCompTipPagId?: number;

    /**
     * @type { number } - nAsigIncMonto
     */
    nAsigIncMonto?: number;

    /**
     * @type { string } - cAsigIncAnotaciones
     */
    cAsigIncAnotaciones?: string;

    /**
     * @type { string } - cCompAccNombre
     */
    cCompAccNombre?: string;

    /**
     * @type { string } - cCompMotAccNombre
     */
    cCompMotAccNombre?: string;

    /**
     * @type { string } - cRegLabNombre
     */
    cRegLabNombre?: string;

    /**
     * @type { string } - cAsigIncMotivoPago
     */
    cAsigIncMotivoPago?: string;

    /**
     * @type { number } - IAsigIncId
     */
    IAsigIncId?: number;

    /**
     * @type { string } - cAsigIncFallecido
     */
    cAsigIncFallecido?: string;

}